+++ src/aui-scheduler/HISTORY.md
@@ -1,2 +1,7 @@
-aui-scheduler
-========
+# AUI Scheduler
+
+> Documentation and test modifications are not included in this changelog. For more details, see [full commit history](https://github.com/liferay/alloy-ui/commits/master/src/aui-scheduler).
+
+## @VERSION@
+
+* [AUI-1370](https://issues.liferay.com/browse/AUI-1370) Add ability to set an event's border color, style, and width